done 4.21

+++ part4/blog/controllers/blogs.js
@@ -45,7 +45,21 @@ blogsRouter.post('/',  async (request, response) => {
 })
 
 blogsRouter.delete('/:id',  async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid'})
+  }
+
+  const user = await User.findById(decodedToken.id)
+
+  if (blogToDelete.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'invalid user'})
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
+
   response.status(204).end()
 })
